Deliverable. Will have fun with this in the future
Hello to anyone at Quadax that reads commit messages! I enjoyed the
coding challenge and hope that we can talk again in the future.
I will be returning to this code from time to time to improve it, so
if you find yourself bored and with free time, please feel free to
return.

I hope it brought at least a smile or two in the process!

Congratulations on the 2022 Best in KLAS! It would be an honor
to earn a place on your team.